refactor: directories structure
