+++ src/client/services/index.ts
@@ -1,0 +1,1 @@
+export { ServerConnection } from './server-connection';
