+++ src/server/connection/index.ts
@@ -1,0 +1,2 @@
+export { ClientConnection } from './client-connection';
+export { ConnectionService } from './connection-service';
